feat: View 분리
