+++ WhimsyLand/Views/ObjectPlacementRealityView.swift
@@ -131,95 +131,3 @@ struct ObjectPlacementRealityView: View {
         }
     }
 }
-
-// To-do 추후 View 따로 리펙토링하기
-struct TooltipView: View {
-    var text: String
-
-    var body: some View {
-        Text(text)
-            .font(.caption)
-            .foregroundColor(.secondary)
-            .multilineTextAlignment(.center)
-            .frame(maxWidth: 220)
-            .padding(.horizontal, 8)
-            .padding(.vertical, 12)
-            .glassBackgroundEffect()
-            .allowsHitTesting(false) // Prevent the tooltip from blocking spatial tap gestures.
-    }
-}
-
-#Preview(windowStyle: .plain) {
-    VStack {
-        TooltipView(text: "Text")
-    }
-}
-
-struct PlacementTooltip: View {
-    var placementState: PlacementState
-
-    var body: some View {
-        if let message {
-            TooltipView(text: message)
-        }
-    }
-
-    var message: String? {
-        // Decide on a message to display, in order of importance.
-        if !placementState.planeToProjectOnFound {
-            return "Point the device at a horizontal surface nearby."
-        }
-        if placementState.collisionDetected {
-            return "The space is occupied."
-        }
-        if !placementState.userPlacedAnObject {
-            return "Tap to place objects."
-        }
-        return nil
-    }
-}
-
-#Preview(windowStyle: .plain) {
-    VStack {
-        PlacementTooltip(placementState: PlacementState())
-        PlacementTooltip(placementState: PlacementState().withPlaneFound())
-        PlacementTooltip(placementState:
-            PlacementState()
-                .withPlaneFound()
-                .withCollisionDetected()
-        )
-    }
-}
-
-private extension PlacementState {
-    func withPlaneFound() -> PlacementState {
-        planeToProjectOnFound = true
-        return self
-    }
-
-    func withCollisionDetected() -> PlacementState {
-        activeCollisions = 1
-        return self
-    }
-}
-
-
-struct DeleteButton: View {
-    var deletionHandler: (() -> Void)?
-
-    var body: some View {
-        Button {
-            if let deletionHandler {
-                deletionHandler()
-            }
-        } label: {
-            Image(systemName: "trash")
-        }
-        .accessibilityLabel("Delete object")
-        .glassBackgroundEffect()
-    }
-}
-
-#Preview(windowStyle: .plain) {
-    DeleteButton()
-}
